Add support for encrypted events in unread thread notifications
